requirements.txt
Pillow
matplotlib 